Struct liquidity_elections::repository::ElectionRepository

pub struct ElectionRepository { /* fields omitted */ }

Methods

impl ElectionRepository

pub fn new(cache_capacity: usize, time_to_live: Duration) -> ElectionRepository

pub async fn create_election<'_, '_, T: DbConnection>(
&'_ self,
election: ElectionInput,
creator_id: &'_ str,
conn: T
) -> Result<Election, DatabaseError>

Create a new election in the database

This inserts a new election given an election input and the ID of the user creating it. This will error if the insert fails for some reason (at this point it should only be due to connection errors).

Arguments

election - The input object with the user input data for the new election

creator_id - The id of the user calling the creation function

conn - The database connection to execute the insert on

Example

let election_input = ElectionInput {
    name: Some("test_name".to_string()),
    description: Some("This is a test description".to_string()),
    choices: Some(vec!["test1".to_string(), "test2".to_string()]),
    ..ElectionInput::default()
};

let result = repository.create_election(election_input, "auth0|test", conn)
    .await.unwrap();

assert_eq!(result.name, "test_name".to_string());

pub async fn update_election<'_, '_, T: DbConnection>(
&'_ self,
id: &'_ Uuid,
input: ElectionInput,
conn: T
) -> Result<Election, DatabaseError>

Update an election in the database

This update an election in the database given an election input and the ID of the election. This will error if the stream doesn't exist or the update fails because of connection errors.

Arguments

id - The id of the election to be updated

election - The input object with the fields to update

conn - The database connection to execute the update on

Example

use liquidity::Uuid;
futures::executor::block_on(async {

let election_input = ElectionInput {
    description: Some("This is a new test description".to_string()),
    ..ElectionInput::default()
};

let id = Uuid::new_v4();
let result = repository.update_election(&id, election_input, conn).await;

assert!(result.is_err())

pub async fn find_election<'_, '_, T: DbConnection>(
&'_ self,
id: &'_ Uuid,
conn: T
) -> Result<Option<Election>, DatabaseError>

Find an election by its id

Arguments

id - The id of the election

conn - The database connection

Returns

The election or the database error if one occurred

Example

let id = Uuid::new_v4();

let election = repository.find_election(&id, conn).await.unwrap();

assert_eq!(election, None);