Function liquidity_elections::resolvers::query::election

pub fn election(id: Uuid, context: &Context) -> Result<Option<Election>, Error>

Fetch an election by id

Arguments

id - The id to look up the election by context - The request context, passed automatically

Permissions Required

view:election

Returns

The election if it exists, None if it doesn't, Error if an issue has occurred

Example

query {
    election(id: "some_uuid") {
        id
        name
        choices
    }
}